Increase max sector limit in Universe

+++ src/common/universe/Universe.cpp
@@ -6,7 +6,7 @@ namespace voxel_game::universe {
 		return it == mObjects.end()? nullptr: it->second.get();
 	}
 
-	std::vector<uint32_t> Universe::getPhysicsObjectsInSector(const glm::ivec3 &pos) {
+	std::vector<uint32_t> Universe::getPhysicsObjectsInSector(const glm::i64vec3 &pos) {
 		const auto it = mSectorPhysicsObjects.find(pos);
 		return it == mSectorPhysicsObjects.end()? std::vector<uint32_t>{}: it->second;
 	}
